requirement.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
distlib==0.3.8
Django==3.2.4
djangorestframework==3.15.1
executing==1.2.0
filelock==3.13.1
jedi==0.19.0
parso==0.8.3
pickleshare==0.7.5
pika==1.3.1
platformdirs==4.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.5
pure-eval==0.2.2
py4j==0.10.9.5
pyflink==1.0
Pygments==2.15.1
pyspark==3.3.2
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
traitlets==5.9.0
tzdata==2024.1
virtualenv==20.25.0
wcwidth==0.2.6